Renamed jag_trainable_models to jag_models.
